Updated TargetFramework and removed certain ItemGroups
Updated the `TargetFramework` in `FrontEnd.csproj` and `QRCode.csproj` from `net6.0` to `net8.0`. Removed the `<ItemGroup>` that included the `wwwroot\img\` folder in `FrontEnd.csproj` and the `<ItemGroup>` that included the `Data\` folder and updated the `Data\WifiData.cry` file in `QRCode.csproj`. No changes were made to the package references.